Add lc0077

+++ lib/combination_iterator.h
@@ -6,10 +6,10 @@
 
 struct CombinationIterator
 {
-    size_t* subset;
-    size_t i;
-    size_t k;
-    size_t n;
+    int* subset;
+    int i;
+    int k;
+    int n;
     bool end;
 };
 
@@ -17,8 +17,8 @@ typedef struct CombinationIterator* CombinationIterator;
 
 void combination_begin(
     CombinationIterator iterator,
-    size_t subset[],
-    size_t n,
-    size_t k);
+    int subset[],
+    int n,
+    int k);
 
 void combination_next(CombinationIterator iterator);
